+++ lib/features/password_manager/dashboard_v2/providers/filter_providers/filter_providers.dart
@@ -1,0 +1,19 @@
+library;
+
+export 'api_keys_filter_provider.dart';
+export 'bank_cards_filter_provider.dart';
+export 'base_filter_provider.dart';
+export 'certificates_filter_provider.dart';
+export 'contacts_filter_provider.dart';
+export 'crypto_wallets_filter_provider.dart';
+export 'documents_filter_provider.dart';
+export 'files_filter_provider.dart';
+export 'identities_filter_provider.dart';
+export 'license_keys_filter_provider.dart';
+export 'notes_filter_provider.dart';
+export 'otp_filter_provider.dart';
+export 'password_filter_provider.dart';
+export 'recovery_codes_filter_provider.dart';
+export 'ssh_keys_filter_provider.dart';
+export 'wifis_filter_provider.dart';
+export 'loyalty_cards_filter_provider.dart';